+++ crates/protocol/src/outgoing.rs
@@ -3,20 +3,16 @@ use serde::Serialize;
 #[derive(Debug, PartialEq, Serialize)]
 #[serde(rename_all = "snake_case", tag = "message")]
 pub enum Message {
-    WriteUrl(WriteUrl),
-    ReadUrl(ReadUrl),
+    /// An access url containing a write session
+    WriteUrl(AccessUrl),
+    /// An access url containing a readonly session
+    ReadUrl(AccessUrl),
     Error(Error),
 }
 
 #[derive(Debug, PartialEq, Serialize)]
 #[serde(rename_all = "snake_case")]
-pub struct WriteUrl {
-    pub url: String,
-}
-
-#[derive(Debug, PartialEq, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub struct ReadUrl {
+pub struct AccessUrl {
     pub url: String,
 }
 
@@ -25,10 +21,14 @@ pub struct ReadUrl {
 pub enum Error {
     /// The requesting user has insufficient permissions for the operation
     InsufficientPermissions,
+    /// The request contains invalid participant ids
+    InvalidParticipantSelection,
     /// Is send when another instance just started initializing and etherpad is not available yet
     CurrentlyInitializing,
     /// The etherpad initialization failed
     FailedInitialization,
+    /// The etherpad is not yet initailized
+    NotInitialized,
 }
 
 #[cfg(test)]
@@ -40,7 +40,7 @@ mod test {
     fn write_url() {
         let json_str = r#"{"message":"write_url","url":"http://localhost/auth_session?sessionID=s.session&padName=protocol&groupID=g.group"}"#;
 
-        let message = Message::WriteUrl(WriteUrl {
+        let message = Message::WriteUrl(AccessUrl {
             url:
                 "http://localhost/auth_session?sessionID=s.session&padName=protocol&groupID=g.group"
                     .into(),
@@ -55,7 +55,7 @@ mod test {
     fn read_url() {
         let json_str = r#"{"message":"read_url","url":"http://localhost:9001/auth_session?sessionID=s.session_id&padName=r.readonly_id"}"#;
 
-        let message = Message::ReadUrl(ReadUrl {
+        let message = Message::ReadUrl(AccessUrl {
             url: "http://localhost:9001/auth_session?sessionID=s.session_id&padName=r.readonly_id"
                 .into(),
         });
@@ -92,6 +92,17 @@ mod test {
         let json_str = r#"{"message":"error","error":"currently_initializing"}"#;
 
         let message = Message::Error(Error::CurrentlyInitializing);
+
+        let string = serde_json::to_string(&message).unwrap();
+
+        assert_eq!(string, json_str)
+    }
+
+    #[test]
+    fn invalid_participant_selection() {
+        let json_str = r#"{"message":"error","error":"invalid_participant_selection"}"#;
+
+        let message = Message::Error(Error::InvalidParticipantSelection);
 
         let string = serde_json::to_string(&message).unwrap();